try_excavate() now throws an error if unable to refuel when needed
This is to ensure that when try_excavate() returns false, the turtle will always be positioned where it was when the function was initially called. As try_refuel() would leave the turtle at the refuel point (presumably further back than this), try_excavate() should not handle this case by simply returning false.